消息提醒前端展示

+++ src/main/java/com/yws/plane/repository/MailRepository.java
@@ -3,5 +3,5 @@ package com.yws.plane.repository;
 import com.yws.plane.entity.Email;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MailRepository extends JpaRepository<Email,Integer> {
+public interface MailRepository extends JpaRepository<Email,Long> {
 }
